refactor(ui): advanced income card refactor
